homework_05/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic_core==2.27.2
pydentic==0.0.1.dev3
pytest==8.3.4
typing_extensions==4.12.2
Jinja2==3.1.5
